requirements.txt
Django==5.2.1
djangorestframework==3.15.0
django-cors-headers==4.1.0
gunicorn==20.1.0
whitenoise==6.5.0